+++ app/Basket/Installation.php
@@ -16,6 +16,18 @@ use Illuminate\Database\Eloquent\Model;
  * Installation Model
  *
  * @author WN
+ * @property int    $id
+ * @property int    $merchant_id
+ * @property string $name
+ * @property true   $active
+ * @property true   $linked
+ * @property int    $ext_id
+ * @property string $ext_name
+ * @property string $ext_return_url
+ * @property string $ext_notification_url
+ * @property string $ext_default_product
+ * @property        $created_at
+ * @property        $updated_at
  * @package App\Basket
  */
 class Installation extends Model