+++ app/Role.php
@@ -1,9 +1,47 @@
 <?php
-
+/*
+ * This file is part of the PayBreak/basket package.
+ *
+ * (c) PayBreak <dev@paybreak.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
 namespace App;
 
 use Zizaco\Entrust\EntrustRole;
-
+use Zizaco\Entrust\Traits\EntrustRoleTrait;
+/**
+ * Class Role
+ *
+ * @author MS
+ * @property int    $id
+ * @property string $name
+ * @property string $display_name
+ * @property string $description
+ * @package App
+ */
 class Role extends EntrustRole
 {
+    use EntrustRoleTrait;
+    /**
+     * The database table used by the model.
+     *
+     * @var string
+     */
+    protected $table = 'roles';
+
+    /**
+     * The attributes that are mass assignable.
+     *
+     * @var array
+     */
+    protected $fillable = ['name', 'display_name', 'description', 'updated_at', 'created_at'];
+
+    /**
+     * The attributes excluded from the model's JSON form.
+     *
+     * @var array
+     */
+    protected $hidden = [];
 }