+++ app/RolePermissions.php
@@ -1,0 +1,43 @@
+<?php
+/*
+ * This file is part of the PayBreak/basket package.
+ *
+ * (c) PayBreak <dev@paybreak.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+namespace App;
+
+use Illuminate\Database\Eloquent\Model;
+/**
+ * Class RolePermissions
+ *
+ * @author MS
+ * @property int    $permission_id
+ * @property int    $role_id
+ * @package App
+ */
+class RolePermissions extends Model  {
+
+    /**
+     * The database table used by the model.
+     *
+     * @var string
+     */
+    protected $table = 'permission_role';
+
+    /**
+     * Attributes that should be mass-assignable.
+     *
+     * @var array
+     */
+    protected $fillable = ['permission_id', 'role_id'];
+
+    /**
+     * The attributes excluded from the model's JSON form.
+     *
+     * @var array
+     */
+    protected $hidden = ['permission_id', 'role_id'];
+}