+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+# 0.2.0
+2015-07-17
+
+## Features
+- Created
+ - `AbstractGateway`
+ - `MerchantGateway`
+ - `MerchantSynchronisationService`
+- Set standards to use API
+- Added Roles CRUD
+- Added Permission Management From Roles
+- Added *Drag And Drop* Pattern For Permissions Management
+- Added SQLite In Memory Configuration For Testing Environment
+
+## Bug Fixes
+- Timestamp Fix For DB Compatibility
+- Fix For Travis CI Encryption Error
+- Fix For CRUD Updating Errors
+
 # 0.1.1
 2015-07-15